refactor: 유틸 클래스 생성
