+++ README.md
@@ -11,14 +11,9 @@
 - [x] Grade 구현
   - 등수에 해당하는 당첨개수와 상금을 가지고 있다.
 
-- [ ] LottoTicket 구현
+- [x] LottoTicket 구현
   - 금액에 맞는 LottoNumbers 를 생성한다.
   - 당첨개수에 맞게 등수를 반환한다.
   - 당첨 번호로 당첨 통계를 만들 수 있다.
   
 - [x] UI구현
-
-- [ ] 유틸클래스 생성
-- [ ] 원시값 포장
-- [ ] getter 자제
-- [ ] 클래스 메소드만 있을 경우 생성자 private
