+++ src/main/java/lotto/controller/LottoMachine.java
@@ -9,6 +9,8 @@ import lotto.utils.IntegerParser;
 import java.util.Collections;
 import java.util.List;
 
+import static lotto.utils.Validator.checkNotNull;
+
 public class LottoMachine {
 
     public static LottoTicket publish(int won) {
@@ -16,6 +18,7 @@ public class LottoMachine {
     }
 
     public static Statistics rank(LottoTicket lottoTicket, String winningNumbers) {
+        checkNotNull(lottoTicket);
         List<Integer> numbers = IntegerParser.listOf(winningNumbers);
         return lottoTicket.rank(LottoNumbers.of(numbers));
     }
