+++ src/main/java/lotto/ui/view/InputView.java
@@ -4,9 +4,10 @@ import java.util.Scanner;
 
 public class InputView {
 
-    private static Scanner SCANNER = new Scanner(System.in);
+    private static final Scanner SCANNER = new Scanner(System.in);
 
-    public static String string;
+    private InputView() {
+    }
 
     public static int getWon() {
         System.out.println("구입금액을 입력해 주세요.");
