+++ src/test/java/lotto/domain/GradeTest.java
@@ -1,6 +1,5 @@
-package lotto;
+package lotto.domain;
 
-import lotto.domain.Grade;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
@@ -12,6 +11,7 @@ class GradeTest {
     @DisplayName("일치하는 개수에 맞게 등급을 반환한다.")
     @ParameterizedTest(name = "[{index}] matchCount: {0}, expected: {1}")
     @CsvSource(value = {
+            "1000, BANG",
             "6, FIRST",
             "5, SECOND",
             "4, THIRD",
