+++ src/test/java/lotto/domain/StatisticsTest.java
@@ -1,8 +1,5 @@
-package lotto;
+package lotto.domain;
 
-import lotto.domain.Dollars;
-import lotto.domain.Grade;
-import lotto.domain.Statistics;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
